Rename IdeToolbarNew to IdeWrapper
related to #360
